:truck: move rabbitmq publisher
